Added Quiz and questioning
Started with questioning section of the system. Users of role type "admin" and "instructor" are the only ones authorized to set a quiz

Currently finished with
create method and delete method

new changes
found a new way to let the insertData to be used across the create and update method. Also found a way to lessen codes in the create method to cause all validation to happen in the validate method